requirements.txt
openai
psycopg2-binary
requests
dotenv
pydantic
google-api-python-client
google_auth_oauthlib
openpyxl